Refactored the code. Now, there is common index.js for Android and IOS
