+++ app/weatherAPI.js
